weather_dashboard/requirements.txt
# Add your library dependencies here
requests
python-dotenv
streamlit